Add admin dashboard, background worker and Docker environment
- apps/admin: Next.js 14 App Router dashboard (TypeScript, Tailwind,
  TanStack Query, Recharts) with overview metrics, user management,
  exercise library, template featuring, AI usage, storage, system health
  and the audit log. Token refresh is transparent; tokens live in
  sessionStorage rather than localStorage. Full dark mode via semantic
  tokens, with loading/empty/error states on every page.
  Admins see photo counts, never photos.
- services/api/app/jobs: arq worker with reminders, weekly summaries,
  goal refresh, stale-session closing, retention cleanup and 30-day
  account purge that also removes stored objects.
- docker-compose.yml: PostgreSQL, Redis, MinIO (+bucket init), API,
  worker and admin. Multi-stage Dockerfiles for both services; the API
  entrypoint waits for the database and applies migrations on start.
- infra/nginx: TLS termination, per-zone rate limits, API/dashboard
  routing; infra/docker: production overlay and deployment notes.
- .env.example documents every variable with safe defaults.

Admin builds clean: 12 routes, lint and typecheck pass.

+++ services/api/app/services/metrics.py
@@ -131,7 +131,7 @@ def estimate_targets(
         "recommended_carbs_g": round(carbs_g),
         "recommended_fat_g": round(fat_g),
         "recommended_fiber_g": round(fiber_g),
-        "recommended_water_ml": int(round(water_ml / 50) * 50),
+        "recommended_water_ml": round(water_ml / 50) * 50,
     }
 
 
@@ -149,7 +149,7 @@ def estimate_session_calories(
     hours = duration_seconds / 3600
     base = 5.0 * weight * hours
     volume_bonus = min(total_volume_kg / 1000, 12) * 4
-    return int(round(base + volume_bonus))
+    return round(base + volume_bonus)
 
 
 def kg_to_lb(value: float) -> float:
